MAINT: Rename sparse module and make them private
