+++ scipy/sparse/_data.py
@@ -8,8 +8,8 @@
 
 import numpy as np
 
-from .base import spmatrix, _ufuncs_with_fixed_point_at_zero
-from .sputils import isscalarlike, validateaxis, matrix
+from ._base import spmatrix, _ufuncs_with_fixed_point_at_zero
+from ._sputils import isscalarlike, validateaxis, matrix
 
 __all__ = []
