+++ scipy/sparse/_generate_sparsetools.py
