+++ scipy/sparse/_sputils.py
